src/backend-py/src/requirements.txt
auto_mix_prep==0.2.0
fastapi==0.122.0
openai==2.8.1
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn==0.38.0
